atualizações 16/09

+++ conta_bancaria/src/main/java/com/senai/conta_bancaria/domain/entity/ContaCorrente.java
@@ -3,11 +3,13 @@ package com.senai.conta_bancaria.domain.entity;
 import jakarta.persistence.Entity;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class ContaCorrente extends Conta{
 
-    private double limite;
+    private BigDecimal limite;
 
-    private double taxa;
+    private BigDecimal taxa;
 }
