+++ conta_bancaria/src/main/java/com/senai/conta_bancaria/domain/entity/ContaPoupanca.java
@@ -3,9 +3,11 @@ package com.senai.conta_bancaria.domain.entity;
 import jakarta.persistence.Entity;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class ContaPoupanca extends Conta{
 
-    private double rendimento;
+    private BigDecimal rendimento;
 }
